move safe_to_server_errors at module level
