+++ argopy/tests/test_fetchers_data_argovis.py
@@ -3,42 +3,16 @@ import xarray as xr
 
 import pytest
 import tempfile
-import warnings
 
 import argopy
 from argopy import DataFetcher as ArgoDataFetcher
 from argopy.errors import (
-    ArgovisServerError,
     CacheFileNotFound,
     FileSystemHasNoCache,
-    DataNotFound,
 )
-from aiohttp.client_exceptions import ServerDisconnectedError
 from argopy.utilities import is_list_of_strings
-from . import requires_connected_argovis
+from . import requires_connected_argovis, safe_to_server_errors
 
-
-def safe_to_server_errors(test_func):
-    """ Test wrapper to make sure we don't fail because of an error from the server, not our Fault ! """
-    def test_wrapper(fix):
-        try:
-            test_func(fix)
-        except ArgovisServerError as e:
-            # Test is passed when something goes wrong because of the argovis server
-            warnings.warn("\nSomething happened on argovis that should not: %s" % str(e.args))
-            pass
-        except DataNotFound as e:
-            # We make sure that data requested by tests are available from API, so this must be a server side error.
-            warnings.warn("\nSomething happened on argovis that should not: %s" % str(e.args))
-            pass
-        except ServerDisconnectedError as e:
-            # We make sure that data requested by tests are available from API, so this must be a server side error.
-            warnings.warn("\nWe were disconnected from server !\n%s" % str(e.args))
-            pass
-        except Exception:
-            raise
-
-    return test_wrapper
 
 
 @requires_connected_argovis
