+++ argopy/tests/test_fetchers_facade_data.py
@@ -8,12 +8,8 @@ import argopy
 from argopy import DataFetcher as ArgoDataFetcher
 from argopy.errors import (
     InvalidFetcherAccessPoint,
-    InvalidFetcher,
-    ErddapServerError,
-    ArgovisServerError,
-    DataNotFound
+    InvalidFetcher
 )
-from aiohttp.client_exceptions import ServerDisconnectedError
 from argopy.utilities import is_list_of_strings
 from . import (
     AVAILABLE_SOURCES,
@@ -21,35 +17,8 @@ from . import (
     requires_connected_erddap_phy,
     requires_localftp,
     requires_connected_argovis,
+    safe_to_server_errors
 )
-
-
-def safe_to_server_errors(test_func):
-    """ Test fixture to make sure we don't fail because of an error from the server, not our Fault ! """
-
-    def test_wrapper(fix):
-        try:
-            test_func(fix)
-        except ErddapServerError as e:
-            # Test is passed when something goes wrong because of the erddap server
-            warnings.warn("\nSomething happened on erddap that should not: %s" % str(e.args))
-            pass
-        except ArgovisServerError as e:
-            # Test is passed when something goes wrong because of the argovis server
-            warnings.warn("\nSomething happened on argovis that should not: %s" % str(e.args))
-            pass
-        except DataNotFound as e:
-            # We make sure that data requested by tests are available from API, so this must be a server side error.
-            warnings.warn("\nSomething happened on server: %s" % str(e.args))
-            pass
-        except ServerDisconnectedError as e:
-            # We make sure that data requested by tests are available from API, so this must be a server side error.
-            warnings.warn("\nWe were disconnected from server !\n%s" % str(e.args))
-            pass
-        except Exception:
-            raise
-
-    return test_wrapper
 
 
 class Test_Facade:
